% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.glmLimits.R
\name{coef.glmLimits}
\alias{coef.glmLimits}
\title{coef.glmLimits function}
\usage{
\method{coef}{glmLimits}(x, ...)
}
\arguments{
\item{x}{an object of class "glmLimits".}
}
\description{
This function shows estimated coeficients estimated with the glmLimits function.
}
\examples{
Limits::coef.glmLimits(x, ...)
}
\keyword{GLM}
\keyword{LOD}
\keyword{bayesian}

