% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.glmLimits.R
\name{confint.glmLimits}
\alias{confint.glmLimits}
\title{confint.glmLimits function}
\usage{
\method{confint}{glmLimits}(x, level = 0.95, alpha = NULL, dec = 4, ...)
}
\arguments{
\item{x}{an object of class "glmLimits".}

\item{level}{the confidence level required.}

\item{alpha}{optional signicance level (ignoring level value).}

\item{dec}{ignore this... (used to match the quantile of the glmLimits object)}
}
\description{
This function compute confidence intervals for all parameters estimated with the glmLimits function.
}
\examples{
Limits::confint.glmLimits(x, level = .95, alpha = NULL, dec = 4, ...)
}
\keyword{GLM}
\keyword{LOD}
\keyword{bayesian}

