% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.glmLimits.R
\name{predict.glmLimits}
\alias{predict.glmLimits}
\title{predict.glmLimits function}
\usage{
\method{predict}{glmLimits}(x, type = c("terms", "binomialresponse",
  "probability"), p = 0.5, ...)
}
\arguments{
\item{x}{an object of class "glmLimits".}

\item{type}{the type of prediction required. The default ("terms") returns fitted values of each term in the model formula on the linear predictor scale. "binomialresponse" returns binary predictions and "probability" return probabilities if family is "binomial"}

\item{p}{probability border value in [0,1] to classify binary outcome.}
}
\description{
This function compute the predictors estimated with the glmLimits function.
}
\examples{
Limits::predict.glmLimits()
}
\keyword{GLM}
\keyword{LOD}
\keyword{bayesian}

