% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.glmLimits.R
\name{print.glmLimits}
\alias{print.glmLimits}
\title{print.glmLimits function}
\usage{
\method{print}{glmLimits}(x, ...)
}
\arguments{
\item{x}{an object of class "glmLimits".}
}
\description{
This function shows a short summary of the glmLimits function results.
}
\examples{
Limits::print.glmLimits()
}
\keyword{GLM}
\keyword{LOD}
\keyword{bayesian}

