% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.rjags.R
\name{print.summary.rjags}
\alias{print.summary.rjags}
\title{print.summary.rjags function}
\usage{
\method{print}{summary.rjags}(object, ...)
}
\arguments{
\item{object}{an object of class "summary.rjags".}
}
\description{
This function shows the summary of the glmLimits function results.
}
\examples{
Limits::print.summary.rjags()
}
\keyword{GLM}
\keyword{LOD}
\keyword{bayesian}

