% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.glmLimits.R
\name{summary.glmLimits}
\alias{summary.glmLimits}
\title{summary.glmLimits function}
\usage{
\method{summary}{glmLimits}(model, ...)
}
\arguments{
\item{model}{the model (of class "jags") extracted from an object of class "glmLimits".}
}
\description{
This function return the summary of the glmLimits function results with class "summary.rjags".
}
\examples{
Limits::summary.glmLimits()
}
\keyword{GLM}
\keyword{LOD}
\keyword{bayesian}

